refactor(kolme): extract notifications consumer module ownership (#1942)
