+++ crates/kamn-core/src/kolme_runtime_commit.rs
@@ -111,6 +111,7 @@ use std::thread;
 use std::time::{Duration, Instant};
 
 mod in_memory_client;
+mod notifications_consumer;
 mod notifications_websocket;
 mod runtime_pipeline;
 
@@ -502,6 +503,7 @@ pub enum KolmeRuntimeCommitOutcome {
 pub type RuntimeCommitLifecycleState = KamnKolmeRuntimeCommitLifecycleState;
 
 pub use in_memory_client::InMemoryKolmeRuntimeCommitClient;
+pub use notifications_consumer::KolmeRuntimeCommitNotificationsConsumer;
 pub use notifications_websocket::{
     KolmeRuntimeCommitWebsocketConnection, KolmeRuntimeCommitWebsocketConnector,
 };
@@ -1368,143 +1370,6 @@ impl<T: KolmeRuntimeCommitBlockFallbackTransport> KolmeRuntimeCommitBlockFallbac
     }
 }
 
-/// Deterministic notifications consumer for Kolme websocket events with bounded reconnect policy.
-pub struct KolmeRuntimeCommitNotificationsConsumer<C>
-where
-    C: KolmeRuntimeCommitNotificationsConnector,
-{
-    notifications_url: String,
-    provider: String,
-    max_reconnect_attempts: u32,
-    connector: C,
-    connection: Option<C::Connection>,
-}
-
-impl<C> KolmeRuntimeCommitNotificationsConsumer<C>
-where
-    C: KolmeRuntimeCommitNotificationsConnector,
-{
-    /// Builds notifications consumer from HTTP base URL and notifications path.
-    pub fn new(
-        base_url: &str,
-        notifications_path: &str,
-        provider: &str,
-        max_reconnect_attempts: u32,
-        connector: C,
-    ) -> Result<Self, KolmeRuntimeCommitError> {
-        if !is_kolme_valid_notifications_provider_input_contract(provider) {
-            return Err(KolmeRuntimeCommitError::InvalidRequest {
-                field: "notifications_provider",
-                reason: "must not be empty",
-            });
-        }
-        if !is_kolme_valid_notifications_reconnect_budget_contract(max_reconnect_attempts) {
-            return Err(KolmeRuntimeCommitError::InvalidRequest {
-                field: "notifications_max_reconnect_attempts",
-                reason: "must be positive",
-            });
-        }
-        let notifications_url =
-            compose_kolme_notifications_websocket_url(base_url, notifications_path)
-                .map_err(|error| KolmeRuntimeCommitProviderError::Unavailable {
-                    reason: error.to_string(),
-                })
-                .map_err(|error| match error {
-                    KolmeRuntimeCommitProviderError::Timeout => {
-                        KolmeRuntimeCommitError::ProviderTransport {
-                            kind: KolmeRuntimeCommitTransportErrorKind::Timeout,
-                            detail: "provider request timed out".to_owned(),
-                        }
-                    }
-                    KolmeRuntimeCommitProviderError::Unavailable { reason } => {
-                        KolmeRuntimeCommitError::ProviderTransport {
-                            kind: KolmeRuntimeCommitTransportErrorKind::Unavailable,
-                            detail: reason,
-                        }
-                    }
-                    KolmeRuntimeCommitProviderError::MalformedResponse { reason } => {
-                        KolmeRuntimeCommitError::ProviderTransport {
-                            kind: KolmeRuntimeCommitTransportErrorKind::MalformedResponse,
-                            detail: reason,
-                        }
-                    }
-                })?;
-        Ok(Self {
-            notifications_url,
-            provider: normalize_kolme_notifications_provider_input_contract(provider).to_owned(),
-            max_reconnect_attempts,
-            connector,
-            connection: None,
-        })
-    }
-
-    /// Reads and parses one notifications event, reconnecting when the stream drops.
-    pub fn next_notification_event(
-        &mut self,
-    ) -> Result<KolmeRuntimeCommitNotificationEvent, KolmeRuntimeCommitProviderError> {
-        let mut reconnect_attempts = 0_u32;
-
-        loop {
-            if self.connection.is_none() {
-                match self.connector.connect(self.notifications_url.as_str()) {
-                    Ok(connection) => self.connection = Some(connection),
-                    Err(_) => {
-                        reconnect_attempts += 1;
-                        if reconnect_attempts >= self.max_reconnect_attempts {
-                            return Err(KolmeRuntimeCommitProviderError::Unavailable {
-                                reason:
-                                    compose_kolme_notifications_reconnect_exhausted_reason_contract(
-                                        self.max_reconnect_attempts,
-                                    ),
-                            });
-                        }
-                        continue;
-                    }
-                }
-            }
-
-            let result = self
-                .connection
-                .as_mut()
-                .expect("connection should exist before read")
-                .read_text_message();
-            match result {
-                Ok(Some(payload)) => {
-                    let event = parse_kolme_notification_event_contract(payload.as_str()).map_err(
-                        |error| KolmeRuntimeCommitProviderError::MalformedResponse {
-                            reason: error.to_string(),
-                        },
-                    )?;
-                    return Ok(event.into());
-                }
-                Ok(None) | Err(_) => {
-                    self.connection = None;
-                    reconnect_attempts += 1;
-                    if reconnect_attempts >= self.max_reconnect_attempts {
-                        return Err(KolmeRuntimeCommitProviderError::Unavailable {
-                            reason: compose_kolme_notifications_reconnect_exhausted_reason_contract(
-                                self.max_reconnect_attempts,
-                            ),
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    /// Reads notification events until one can be mapped to a commit receipt.
-    pub fn next_commit_receipt(
-        &mut self,
-    ) -> Result<KolmeRuntimeCommitProviderReceipt, KolmeRuntimeCommitProviderError> {
-        loop {
-            let event = self.next_notification_event()?;
-            if let Some(receipt) = event.to_provider_receipt(self.provider.as_str()) {
-                return Ok(receipt);
-            }
-        }
-    }
-}
-
 /// Fork-profile finality resolver that composes notifications and block fallback lookups.
 pub struct KolmeRuntimeCommitForkFinalityResolver<C, T>
 where
@@ -1557,7 +1422,7 @@ where
                 }
                 _ => {
                     let receipt = event
-                        .to_provider_receipt(self.notifications_consumer.provider.as_str())
+                        .to_provider_receipt(self.notifications_consumer.provider())
                         .ok_or_else(|| KolmeRuntimeCommitProviderError::MalformedResponse {
                             reason: "notification event did not carry receipt data".to_owned(),
                         })?;
