+++ crates/kamn-core/src/kolme_runtime_commit/notifications_consumer.rs
@@ -1,0 +1,154 @@
+//! Deterministic notifications consumer for Kolme websocket event streams.
+
+use super::{
+    compose_kolme_notifications_reconnect_exhausted_reason_contract,
+    compose_kolme_notifications_websocket_url,
+    is_kolme_valid_notifications_provider_input_contract,
+    is_kolme_valid_notifications_reconnect_budget_contract,
+    normalize_kolme_notifications_provider_input_contract, parse_kolme_notification_event_contract,
+    KolmeRuntimeCommitError, KolmeRuntimeCommitNotificationEvent,
+    KolmeRuntimeCommitNotificationsConnection, KolmeRuntimeCommitNotificationsConnector,
+    KolmeRuntimeCommitProviderError, KolmeRuntimeCommitProviderReceipt,
+    KolmeRuntimeCommitTransportErrorKind,
+};
+
+/// Deterministic notifications consumer for Kolme websocket events with bounded reconnect policy.
+pub struct KolmeRuntimeCommitNotificationsConsumer<C>
+where
+    C: KolmeRuntimeCommitNotificationsConnector,
+{
+    notifications_url: String,
+    provider: String,
+    max_reconnect_attempts: u32,
+    connector: C,
+    connection: Option<C::Connection>,
+}
+
+impl<C> KolmeRuntimeCommitNotificationsConsumer<C>
+where
+    C: KolmeRuntimeCommitNotificationsConnector,
+{
+    /// Builds notifications consumer from HTTP base URL and notifications path.
+    pub fn new(
+        base_url: &str,
+        notifications_path: &str,
+        provider: &str,
+        max_reconnect_attempts: u32,
+        connector: C,
+    ) -> Result<Self, KolmeRuntimeCommitError> {
+        if !is_kolme_valid_notifications_provider_input_contract(provider) {
+            return Err(KolmeRuntimeCommitError::InvalidRequest {
+                field: "notifications_provider",
+                reason: "must not be empty",
+            });
+        }
+        if !is_kolme_valid_notifications_reconnect_budget_contract(max_reconnect_attempts) {
+            return Err(KolmeRuntimeCommitError::InvalidRequest {
+                field: "notifications_max_reconnect_attempts",
+                reason: "must be positive",
+            });
+        }
+        let notifications_url =
+            compose_kolme_notifications_websocket_url(base_url, notifications_path)
+                .map_err(|error| KolmeRuntimeCommitProviderError::Unavailable {
+                    reason: error.to_string(),
+                })
+                .map_err(|error| match error {
+                    KolmeRuntimeCommitProviderError::Timeout => {
+                        KolmeRuntimeCommitError::ProviderTransport {
+                            kind: KolmeRuntimeCommitTransportErrorKind::Timeout,
+                            detail: "provider request timed out".to_owned(),
+                        }
+                    }
+                    KolmeRuntimeCommitProviderError::Unavailable { reason } => {
+                        KolmeRuntimeCommitError::ProviderTransport {
+                            kind: KolmeRuntimeCommitTransportErrorKind::Unavailable,
+                            detail: reason,
+                        }
+                    }
+                    KolmeRuntimeCommitProviderError::MalformedResponse { reason } => {
+                        KolmeRuntimeCommitError::ProviderTransport {
+                            kind: KolmeRuntimeCommitTransportErrorKind::MalformedResponse,
+                            detail: reason,
+                        }
+                    }
+                })?;
+        Ok(Self {
+            notifications_url,
+            provider: normalize_kolme_notifications_provider_input_contract(provider).to_owned(),
+            max_reconnect_attempts,
+            connector,
+            connection: None,
+        })
+    }
+
+    pub(crate) fn provider(&self) -> &str {
+        self.provider.as_str()
+    }
+
+    /// Reads and parses one notifications event, reconnecting when the stream drops.
+    pub fn next_notification_event(
+        &mut self,
+    ) -> Result<KolmeRuntimeCommitNotificationEvent, KolmeRuntimeCommitProviderError> {
+        let mut reconnect_attempts = 0_u32;
+
+        loop {
+            if self.connection.is_none() {
+                match self.connector.connect(self.notifications_url.as_str()) {
+                    Ok(connection) => self.connection = Some(connection),
+                    Err(_) => {
+                        reconnect_attempts += 1;
+                        if reconnect_attempts >= self.max_reconnect_attempts {
+                            return Err(KolmeRuntimeCommitProviderError::Unavailable {
+                                reason:
+                                    compose_kolme_notifications_reconnect_exhausted_reason_contract(
+                                        self.max_reconnect_attempts,
+                                    ),
+                            });
+                        }
+                        continue;
+                    }
+                }
+            }
+
+            let result = self
+                .connection
+                .as_mut()
+                .expect("connection should exist before read")
+                .read_text_message();
+            match result {
+                Ok(Some(payload)) => {
+                    let event = parse_kolme_notification_event_contract(payload.as_str()).map_err(
+                        |error| KolmeRuntimeCommitProviderError::MalformedResponse {
+                            reason: error.to_string(),
+                        },
+                    )?;
+                    return Ok(event.into());
+                }
+                Ok(None) | Err(_) => {
+                    self.connection = None;
+                    reconnect_attempts += 1;
+                    if reconnect_attempts >= self.max_reconnect_attempts {
+                        return Err(KolmeRuntimeCommitProviderError::Unavailable {
+                            reason: compose_kolme_notifications_reconnect_exhausted_reason_contract(
+                                self.max_reconnect_attempts,
+                            ),
+                        });
+                    }
+                }
+            }
+        }
+    }
+
+    /// Reads notification events until one can be mapped to a commit receipt.
+    pub fn next_commit_receipt(
+        &mut self,
+    ) -> Result<KolmeRuntimeCommitProviderReceipt, KolmeRuntimeCommitProviderError> {
+        loop {
+            let event = self.next_notification_event()?;
+            if let Some(receipt) = event.to_provider_receipt(self.provider.as_str()) {
+                return Ok(receipt);
+            }
+        }
+    }
+}
