+++ crates/kamn-core/tests/kolme_runtime_commit_extraction_plan_docs.rs
@@ -70,6 +70,7 @@ fn regression_requires_phase_gates_and_validation_matrix_markers() {
     assert!(DOC.contains("#1936"));
     assert!(DOC.contains("#1938"));
     assert!(DOC.contains("#1940"));
+    assert!(DOC.contains("#1942"));
     assert!(DOC.contains("## Validation Matrix"));
     assert!(DOC.contains("Regression: #1814"));
 }
